requirements.txt
# Install the following dependencies

# Package                Version
# ---------------------- --------
blinker==1.8.2
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
flask==3.0.3
Flask-Mail==0.9.1
flask-session==0.8.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
mysql-connector-python==8.4.0
mysqlclient==2.2.4
PyJWT==2.8.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.3
zipp==3.18.2
gunicorn==22.0.0 # for production
reportlab== 4.2.0

# How to install all requirements
# pip install -r requirements.txt


# delete __pycache__
# find . -type d -name "__pycache__" -exec rm -r {} +


# sudo apt install python3-mysqldb # for mysqlclient
# npm install @googlemaps/js-api-loader
#gunicorn service
# [Unit]
# Description=Gunicorn daemon/instance to serve ePermit Web App
# After=network.target

# [Service]
# User=ubuntu
# Group=www-data
# WorkingDirectory=/home/ubuntu/ePermit-KLF
# Environment="PATH=/home/ubuntu/ePermit-KLF/epermit/bin:/usr/bin"
# ExecStart=/bin/bash -c 'source /home/ubuntu/ePermit-KLF/epermit/bin/activate && gunicorn --workers 3 --bind 0.0.0.0:5003 web_flask.app:app --error-logfile /tmp/epermit-error.log --access-logfile /tmp/epermit-access.log'

# [Install]
# WantedBy=multi-user.target






# deployment server requirements
# Package                Version
# ---------------------- --------------------
attrs==19.3.0
Automat==0.8.0
blinker                1.8.2
cachelib               0.13.0
certifi                2024.2.2
cffi                   1.16.0
chardet                3.0.4
charset-normalizer     3.3.2
click                  8.1.7
cloud-init             23.1.2
colorama               0.4.3
command-not-found      0.3
configobj              5.0.6
constantly             15.1.0
cryptography           42.0.7
dbus-python            1.2.16
distro                 1.4.0
distro-info            0.23ubuntu1
entrypoints            0.3
flask                  3.0.3
Flask-Mail             0.9.1
flask-session          0.8.0
flask-sqlalchemy       3.1.1
greenlet               3.0.3
gunicorn               20.0.4
httplib2               0.14.0
hyperlink              19.0.0
idna                   3.7
importlib-metadata     7.1.0
incremental            16.10.1
itsdangerous           2.2.0
jinja2                 3.1.4
jsonpatch              1.22
jsonpointer            2.0
jsonschema             3.2.0
keyring                18.0.1
language-selector      0.1
launchpadlib           1.10.13
lazr.restfulclient     0.14.2
lazr.uri               1.0.3
MarkupSafe             2.1.5
more-itertools         4.2.0
msgspec                0.18.6
mysqlclient            1.4.4
netifaces              0.10.4
oauthlib               3.1.0
pexpect                4.6.0
pip                    20.0.2
pyasn1                 0.4.2
pyasn1-modules         0.2.1
pycparser              2.22
PyGObject              3.36.0
PyHamcrest             1.9.0
PyJWT                  2.8.0
pymacaroons            0.13.0
PyNaCl                 1.3.0
pyOpenSSL              24.1.0
pyrsistent             0.15.5
pyserial               3.4
python-apt             2.0.0+ubuntu0.20.4.7
python-debian          0.1.36ubuntu1
python-dotenv          1.0.1
PyYAML                 5.3.1
requests               2.31.0
requests-unixsocket    0.2.0
SecretStorage          2.3.1
service-identity       18.1.0
setuptools             45.2.0
simplejson             3.16.0
six                    1.14.0
sos                    4.3
SQLAlchemy             2.0.30
ssh-import-id          5.10
systemd-python         234
Twisted                18.9.0
typing-extensions      4.11.0
ubuntu-advantage-tools 27.9
ufw                    0.36
unattended-upgrades    0.1
urllib3                2.2.1
wadllib                1.3.3
werkzeug               3.0.3
wheel                  0.34.2
zipp                   3.18.2
zope.interface